refactor: broke into smaller components and removed unnecessary components
